Events and Posts working
Might be inefficient and costly. It used to run for free but cost a
little this month due to datastore writes.

+++ src/main/java/ie/sortons/events/shared/WallPost.java
@@ -24,7 +24,7 @@ public class WallPost implements JsonSerializable {
 	public String postId;
 
 	@Index
-	public int date;
+	public Long date;
 
 	public String url;
 
@@ -37,7 +37,7 @@ public class WallPost implements JsonSerializable {
 	public WallPost() {
 	}
 
-	public WallPost(Long clientId, String postId, int date, String url) {
+	public WallPost(Long clientId, String postId, Long date, String url) {
 		this.clientId = clientId;
 		this.postId = postId;
 		this.date = date;
